Cleanup.

+++ threedigrid_builder/tests/test_quadtree.py
@@ -12,7 +12,7 @@ import pytest
 def subgrid_meta():
     width = 20
     height = 16
-    mask = np.ones((width, height), dtype=np.int32, order="F")
+    mask = np.ones((width, height), dtype=np.int16, order="F")
     mask[15:, :] = 0
     return {
         "pixel_size": 0.5,
